把热键合并到一个文件
